requirements.txt
# News/Content Collection
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=5.0.0

# YouTube
youtube-transcript-api>=0.6.0

# AI Analysis
google-genai>=1.0.0

# PDF Processing & OCR
PyMuPDF>=1.23.0
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=10.0.0

# Discord
discord-webhook>=1.3.0

# Configuration/Utils
python-dotenv>=1.0.0
pyyaml>=6.0

# Date/Time
python-dateutil>=2.8.0
pytz>=2024.1
holidays>=0.40
